LayoutTests/fast/dom/shadow/import-shadowroot.html

<!DOCTYPE html>
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
<keygen id='K'/>
<div id='X'></div>
<pre id="console">
This tests importNode() with bound elements.

</pre>
<script>
    if (window.layoutTestController) {
        layoutTestController.dumpAsText();

        var dstDoc = new DOMParser().parseFromString('<div/>', 'text/xml');
        shouldBeNonNull('dstDoc');

        var srcKeygen = document.getElementById('K');
        shouldBeNonNull('srcKeygen');
        var srcShadow = layoutTestController.shadowRoot(srcKeygen);
        shouldBeNonNull('srcShadow');
        shouldBeNonNull('srcShadow.firstChild');

        var importedKeygen = dstDoc.importNode(srcKeygen);
        shouldBeNonNull('importedKeygen');
        shouldBeTrue('srcKeygen !== importedKeygen');

        var importedShadow = layoutTestController.shadowRoot(importedKeygen);
        shouldBeNonNull('importedShadow');
        shouldBeNonNull('importedShadow.firstChild');
    }

    var successfullyParsed = true;
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
