LayoutTests/fast/dom/shadow/import-shadowroot-expected.txt


This tests importNode() with bound elements.

PASS dstDoc is non-null.
PASS srcKeygen is non-null.
PASS srcShadow is non-null.
PASS srcShadow.firstChild is non-null.
PASS importedKeygen is non-null.
PASS srcKeygen !== importedKeygen is true
PASS importedShadow is non-null.
PASS importedShadow.firstChild is non-null.
PASS successfullyParsed is true

TEST COMPLETE
